+++ naplib/naplab/process_ieeg.py
@@ -328,7 +328,6 @@ def process_ieeg(
     else:
         raise ValueError(f'Unknown string rereference_grid mode: {rereference_grid}')
 
-
     if rereference_grid is not None:
         logger.info(f'Performing common rereferencing using "{rereference_method}" method...')
         if store_reference: