+++ naplib/utils/__init__.py
@@ -1,4 +1,3 @@
 from .argchecking import _parse_outstruct_args
 
 __all__ = ['_parse_outstruct_args']
-