+++ naplib/visualization/__init__.py
@@ -1,4 +1,5 @@
 from .plots import shaded_error_plot, hierarchical_cluster_plot, strf_plot, freq_response, kde_plot
+from .eeg_cap_positions import eeg_locs
 
-__all__  = ['shaded_error_plot', 'hierarchical_cluster_plot', 'strf_plot', 'freq_response', 'kde_plot']
+__all__  = ['eeg_locs', 'shaded_error_plot', 'hierarchical_cluster_plot', 'strf_plot', 'freq_response', 'kde_plot']