+++ docs/changelog.rst
@@ -31,6 +31,7 @@ Change tags (adopted from `sklearn <https://scikit-learn.org/stable/whats_new/v0
 Version 1.1.0 (In development)
 ------------------------------
 - |Enhancement| : Added a ``pre_post`` argument to ``stats.responsive_ttest`` to allow greater flexibility to how the responsiveness test is conducted. Also fixed a minor issue with how the test was being computed.
+- |Enhancement| : Added the function ``visualization.eeg_locs`` to load the EEG channel locations which can be used with MNE ``plot_topomap`` to plot EEG channel data on a scalp.
 
 Version 1.0.0
 -------------